Stop using emacsclient, for now
Aside from subtle issues with using the client in multiple tabs, it's recently
started hanging on startup (I think it might have something to do with
auto-save files, but I'm not sure).  With my new Mac with a solid state drive,
it's fast enough to open and close regular Emacs on a regular basis, so
perhaps this won't be a big deal after all.