Let's try using mmm-mode to render docstrings as rst for a while
The main issue here is the coloring of the docstring text. I figured out how
to make it not green (the default), but the only other color it can be it
seems is bold white.

Probably what I'll have to do is create a submode of rst-mode with a custom
face.