Don't set indent-line-function to insert-tab
This makes auto-fill in TeX documents indent the filled line.

This doesn't seem to break anything, at least not Python. It seems that many
modes define their own indent-line-function anyway. This change was just one
of many that I tossed in to try to make indentation use spaces instead of
tabs, and it's very probably not needed.